Create: scroll search results & scroll control panel
User can scroll when viewing or editing on the control panel. A maximum
of 10 results are shown on the search autocomplete at any given time
(user can scroll or navigate with keyboard for more results). Dates on
control panel are now being correct for different locales. Problems
with new node colors were fixed.
- New features
- Enhancements
- Minor bug fixes